Move clipboard import logic to WalletHeader
