+++ components/WalletHeader.tsx
@@ -1,13 +1,19 @@
 import React from 'react';
 import { Button, Header } from 'react-native-elements';
 import { TouchableOpacity, View } from 'react-native';
+import { inject, observer } from 'mobx-react';
+import Clipboard from '@react-native-clipboard/clipboard';
+
+import SettingsStore from '../stores/SettingsStore';
 
 import LoadingIndicator from '../components/LoadingIndicator';
 import NodeIdenticon from '../components/NodeIdenticon';
+
+import { isClipboardValue } from '../utils/handleAnything';
 import { themeColor } from '../utils/ThemeUtils';
 
 import Contact from '../assets/images/SVG/Mascot contact.svg';
-import Clipboard from '../assets/images/SVG/Clipboard.svg';
+import ClipboardSVG from '../assets/images/SVG/Clipboard.svg';
 import Scan from '../assets/images/SVG/Scan.svg';
 
 import stores from '../stores/Stores';
@@ -48,75 +54,109 @@ const ClipboardBadge = ({
     <TouchableOpacity
         onPress={() => navigation.navigate('Send', { destination: clipboard })}
     >
-        <Clipboard fill={themeColor('text')} width="27" height="27" />
+        <ClipboardSVG fill={themeColor('text')} width="27" height="27" />
     </TouchableOpacity>
 );
 
-export function WalletHeader({
-    navigation,
-    SettingsStore,
-    loading = false,
-    title,
-    channels = false,
-    clipboard
-}: {
+interface WalletHeaderProps {
+    SettingsStore: SettingsStore;
     navigation: any;
-    SettingsStore: any;
-    loading?: boolean;
-    title?: string;
-    channels?: boolean;
-    clipboard?: string;
-}) {
-    const { settings } = SettingsStore;
-    const multipleNodes: boolean =
-        settings && settings.nodes && settings.nodes.length > 1;
-    const selectedNode: any =
-        (settings &&
-            settings.nodes &&
-            settings.nodes[settings.selectedNode || 0]) ||
-        null;
+    loading: boolean;
+    title: string;
+    channels: boolean;
+}
 
-    const SettingsButton = () => (
-        <TouchableOpacity onPress={() => navigation.navigate('Settings')}>
-            {multipleNodes ? (
-                <NodeIdenticon selectedNode={selectedNode} width={30} rounded />
-            ) : (
-                <Contact width={30} />
-            )}
-        </TouchableOpacity>
-    );
+interface WalletHeaderState {
+    clipboard: string;
+}
 
-    return (
-        <Header
-            leftComponent={loading ? undefined : <SettingsButton />}
-            centerComponent={title ? <Body bold>{title}</Body> : null}
-            rightComponent={
-                channels ? (
-                    <OpenChannelButton navigation={navigation} />
-                ) : (
-                    <View style={{ flex: 1, flexDirection: 'row' }}>
-                        {(stores.balanceStore.loadingBlockchainBalance ||
-                            stores.balanceStore.loadingLightningBalance) && (
-                            <LoadingIndicator size={80} />
-                        )}
-                        {!!clipboard && (
-                            <View style={{ marginTop: 14, marginRight: 20 }}>
-                                <ClipboardBadge
-                                    navigation={navigation}
-                                    clipboard={clipboard}
-                                />
-                            </View>
-                        )}
-                        <View style={{ marginTop: 15 }}>
-                            <ScanBadge navigation={navigation} />
-                        </View>
-                    </View>
-                )
+@inject('SettingsStore')
+@observer
+export default class WalletHeader extends React.Component<
+    WalletHeaderProps,
+    WalletHeaderState
+> {
+    state = {
+        clipboard: ''
+    };
+
+    async UNSAFE_componentWillMount() {
+        const { SettingsStore } = this.props;
+        const { settings } = SettingsStore;
+
+        if (settings.privacy && settings.privacy.clipboard) {
+            const clipboard = await Clipboard.getString();
+
+            if (!!clipboard && isClipboardValue(clipboard)) {
+                this.setState({
+                    clipboard
+                });
             }
-            backgroundColor="transparent"
-            containerStyle={{
-                borderBottomWidth: 0
-            }}
-        />
-    );
+        }
+    }
+
+    render() {
+        const { clipboard } = this.state;
+        const { navigation, loading, title, channels, SettingsStore } =
+            this.props;
+        const { settings } = SettingsStore;
+        const multipleNodes: boolean =
+            (settings && settings.nodes && settings.nodes.length > 1) || false;
+        const selectedNode: any =
+            (settings &&
+                settings.nodes &&
+                settings.nodes[settings.selectedNode || 0]) ||
+            null;
+
+        const SettingsButton = () => (
+            <TouchableOpacity onPress={() => navigation.navigate('Settings')}>
+                {multipleNodes ? (
+                    <NodeIdenticon
+                        selectedNode={selectedNode}
+                        width={30}
+                        rounded
+                    />
+                ) : (
+                    <Contact width={30} />
+                )}
+            </TouchableOpacity>
+        );
+
+        return (
+            <Header
+                leftComponent={loading ? undefined : <SettingsButton />}
+                centerComponent={title ? <Body bold>{title}</Body> : null}
+                rightComponent={
+                    channels ? (
+                        <OpenChannelButton navigation={navigation} />
+                    ) : (
+                        <View style={{ flex: 1, flexDirection: 'row' }}>
+                            {(stores.balanceStore.loadingBlockchainBalance ||
+                                stores.balanceStore
+                                    .loadingLightningBalance) && (
+                                <LoadingIndicator size={80} />
+                            )}
+                            {!!clipboard && (
+                                <View
+                                    style={{ marginTop: 14, marginRight: 20 }}
+                                >
+                                    <ClipboardBadge
+                                        navigation={navigation}
+                                        clipboard={clipboard}
+                                    />
+                                </View>
+                            )}
+                            <View style={{ marginTop: 15 }}>
+                                <ScanBadge navigation={navigation} />
+                            </View>
+                        </View>
+                    )
+                }
+                backgroundColor="transparent"
+                containerStyle={{
+                    borderBottomWidth: 0
+                }}
+            />
+        );
+    }
 }
